requirements.txt
# Crawler Service Dependencies
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
asyncio==3.4.3
playwright==1.40.0